fix(proto): address review — proto docs, portable size check, test cleanup
- Add field-level doc comments to vm_service.proto referencing ADR-0006 §2
- Cast len(payload) to uint64 before comparing with math.MaxUint32 for
  32-bit portability
- Consolidate framing tests into a single table-driven TestFraming
- Add nil-payload round-trip test case

Partially generated by AI.

+++ pkg/vsockframing/framing.go
@@ -9,7 +9,7 @@ import (
 
 // WriteFrame writes a length-prefixed frame: [4-byte big-endian uint32 length][payload].
 func WriteFrame(w io.Writer, payload []byte) error {
-	if len(payload) > math.MaxUint32 {
+	if uint64(len(payload)) > math.MaxUint32 {
 		return fmt.Errorf("frame payload too large: %d bytes", len(payload))
 	}
 	length := uint32(len(payload))
